second_task/requirements.txt
numpy>=1.20.0
pillow>=8.0.0
manim>=0.17.2
reportlab>=3.6.0
scipy>=1.7.0 